Conditional compilation of derive_inspector_reflection + pull out common types
