+++ crates/gpui_macros/Cargo.toml
@@ -8,6 +8,9 @@ license = "Apache-2.0"
 [lints]
 workspace = true
 
+[features]
+inspector = []
+
 [lib]
 path = "src/gpui_macros.rs"
 proc-macro = true
@@ -21,4 +24,4 @@ syn.workspace = true
 workspace-hack.workspace = true
 
 [dev-dependencies]
-gpui.workspace = true
+gpui = { workspace = true, features = ["inspector"] }
